requirements.txt
fastapi 
uvicorn 
jinja2
python-frontmatter